Software Project
in this project, I implemented a java code that detects and groups node clusters together. this graph project can be helpful in many scenarios in which there needs to be a grouping of different items to groups that best represents them. such as towns on a map, items in a supermarket...